Hacky but end-to-end bounding box search
In order to dodge the question of how best to receive a set of co-ordinates or
block of GeoJSON in the query I set this up to take a mapit area ID, get the
geoJSON from Mapit and use that to make the query. I've now seen that working
end-to-end and returning results.

Next step will probably be to decouple this from mapit and of course to begin
to add tests.